+++ frontend/app/nextjs-blog/components/todo/todo.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import ReactDOM from "react-dom";
-import createReactClass from "create-react-class";
+import TodoList from './todoList';
+import TodoForm from './todoForm';
 
-var TodoBox = createReactClass({
-	getInitialState: function () {
+export default function TodoBox() {
+	const getInitialState = () => {
 		return {
 			data: [
 				{"id":"00001","task":"Wake up","complete":"false"},
@@ -11,26 +11,30 @@ var TodoBox = createReactClass({
         {"id":"00003","task":"Go to work","complete":"false"}
 			]
 		};
-	},
-	generateId: function () {
+	}
+
+	const generateId = () =>  {
 		return Math.floor(Math.random()*90000) + 10000;
-	},
-	handleNodeRemoval: function (nodeId) {
+	}
+
+	const handleNodeRemoval = (nodeId) => {
 		var data = this.state.data;
 		data = data.filter(function (el) {
 			return el.id !== nodeId;
 		});
 		this.setState({data});
 		return;
-	},
-	handleSubmit: function (task) {
+	}
+
+	const handleSubmit = (task) => {
 		var data = this.state.data;
-		var id = this.generateId().toString();
+		var id = generateId().toString();
 		var complete = 'false';
 		data = data.concat([{id, task, complete}]);
 		this.setState({data});
-	},
-	handleToggleComplete: function (nodeId) {
+	}
+
+	const handleToggleComplete = (nodeId) => {
 		var data = this.state.data;
 		for (var i in data) {
 			if (data[i].id == nodeId) {
@@ -40,104 +44,14 @@ var TodoBox = createReactClass({
 		}
 		this.setState({data});
 		return;
-	},
-	render: function() {
-		return (
-			<div className="well">
-				<h1 className="vert-offset-top-0">To do:</h1>
-				<TodoList data={this.state.data} removeNode={this.handleNodeRemoval} toggleComplete={this.handleToggleComplete} />
-				<TodoForm onTaskSubmit={this.handleSubmit} />
-			</div>
-		);
 	}
-});
 
-var TodoList = createReactClass({
-	removeNode: function (nodeId) {
-		this.props.removeNode(nodeId);
-		return;
-	},
-	toggleComplete: function (nodeId) {
-		this.props.toggleComplete(nodeId);
-		return;
-	},
-	render: function() {
-		var listNodes = this.props.data.map(function (listItem) {
-			return (
-				<TodoItem key={listItem.id} nodeId={listItem.id} task={listItem.task} complete={listItem.complete} removeNode={this.removeNode} toggleComplete={this.toggleComplete} />
-			);
-		},this);
-		return (
-			<ul className="list-group">
-				{listNodes}
-			</ul>
-		);
-	}
-});
+  return (
+    <div className="well">
+      <h1 className="vert-offset-top-0">To do:</h1>
+      <TodoList data={this.state.data} removeNode={this.handleNodeRemoval} toggleComplete={this.handleToggleComplete} />
+      <TodoForm onTaskSubmit={this.handleSubmit} />
+    </div>
+  );
+};
 
-var TodoItem = createReactClass({
-	removeNode: function (e) {
-		e.preventDefault();
-		this.props.removeNode(this.props.nodeId);
-		return;
-	},
-	toggleComplete: function (e) {
-		e.preventDefault();
-		this.props.toggleComplete(this.props.nodeId);
-		return;
-	},
-	updateClass: function () {
-		
-	},
-	render: function() {
-		var classes = 'list-group-item clearfix';
-		if (this.props.complete === 'true') {
-			classes = classes + ' list-group-item-success';
-		}
-		return (
-			<li className={classes}>
-				{this.props.task}
-				<div className="pull-right" role="group">
-					<button type="button" className="btn btn-xs btn-success img-circle" onClick={this.toggleComplete}>&#x2713;</button> <button type="button" className="btn btn-xs btn-danger img-circle" onClick={this.removeNode}>&#xff38;</button>
-				</div>
-			</li>
-		);
-	}
-});
-
-var TodoForm = createReactClass({
-	doSubmit: function (e) {
-		e.preventDefault();
-		var task = ReactDOM.findDOMNode(this.refs.task).value.trim();
-		if (!task) {
-			return;
-		}
-		this.props.onTaskSubmit(task);
-		ReactDOM.findDOMNode(this.refs.task).value = '';
-		return;
-	},
-	render: function() {
-		return (
-			<div className="commentForm vert-offset-top-2">
-				<hr />
-				<div className="clearfix">
-					<form className="todoForm form-horizontal" onSubmit={this.doSubmit}>
-						<div className="form-group">
-							<label htmlFor="task" className="col-md-2 control-label">Task</label>
-							<div className="col-md-10">
-								<input type="text" id="task" ref="task" className="form-control" placeholder="What do you need to do?" />
-							</div>
-						</div>
-						<div className="row">
-							<div className="col-md-10 col-md-offset-2 text-right">
-								<input type="submit" value="Save Item" className="btn btn-primary" />
-							</div>
-						</div>
-					</form>
-				</div>
-			</div>
-		);
-	}
-});
-
-export default TodoBox;
